IMP: Ported to Win64. FIX: Fixed 64bit size_t and unsigned int typo problems. FIX: Fixed many 64bit static analysis problem.
svn path=/trunk/; revision=758

+++ Source/ZEEngine/ZEGraphics/ZECanvas.cpp
@@ -833,7 +833,7 @@ ZEStaticVertexBuffer* ZECanvas::CreateStaticVertexBuffer()
 	return Buffer;
 }
 
-unsigned int ZECanvas::GetBufferSize()
+size_t ZECanvas::GetBufferSize()
 {
 	return Vertices.GetCount() * sizeof(ZECanvasVertex);
 }
